Switch to building macOS packages on CircleCI
It comes as no surprise that we would eventually run into the Travis CI
time limit for our macOS builds. Unfortunately it seems to timeout when
fixing and inspecting the linkages. In order to give the builds a bit
more time, switch to building on CircleCI's macOS workers, which have a
longer time limit.

ref: https://travis-ci.org/conda-forge/llvmdev-feedstock/builds/412780833