Handle new 10.0 InjectTLIMappings in intel SVML patch
New file InjectTLIMappings also needs to be patched by
intel-D47188-svml-VF.patch. Decided to try leaving the svmlMangle()
inside TLI::getVectorizedFunction() because it is only adding a
'_ha' suffix to the function name when it 'isFast()'.  The
mangling done by the new mangleTLIName() should be able to build on
top of that.