expect fastmath entrypoints in add-TLI-mappings.ll
with adjusted intel-D47188-svml-VF.patch that compiles, one of the
new tests in LLVM 10.0.0 fails because it expects SVML function
entrypoints that do not use the 'fastmath' '_ha' suffix.  However,
the beforementioned patch causes 'fastmath' entrypoints to be used
in this test.

Modifying the new test seems correct to me but I'm leaving it as a
separate patch so that it is easier for others to observe the failure
and notice this fundamentally new patch we need to make for 10.0.0
and beyond.

New patchfile is:
recipe/patches/expect-fastmath-entrypoints-in-add-TLI-mappings.ll.patch